modularize the tests
